+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [3.22.0] - 2021-10-08
+### Added
+- Ability for user to use TAB key on a dropdown list to select an item and
+  move to next form control.
+- Cursor will be focused on the input control of the newly added line after
+  user hits "add resource type" or "add search criterion" buttons.
+
 ## [3.21.0] - 2021-10-05
 ### Changed
 - Simplified pull data step search criteria, keeping only observation code